test: ensure fetch mocks are cleared between tests
For some reason this test started failing with `Adding route with same name or matcher as existing route.`. Fixing that meant adding the `fetchMock.reset()` call, but that revealed an issue where `node-fetch` is required by `fetch-mock` but it does not actually depend on it. So this adds `node-fetch` as a dependency.